This is a cleanup revision.  The compiler was giving warnings about redundant casts, and so I fixed my mistakes and removed each of the offending casts.
git-svn-id: https://bbl.ancl.hawaii.edu/projects/bbl/trunk@667 9e959af2-1677-4b95-8350-622619c7fe3b

+++ src/java/edu/hawaii/soest/kilonalu/adam/AdamParser.java
@@ -744,7 +744,7 @@ public class AdamParser {
     float voltage =
     ( this.voltageSenseRange * ( channelData / this.voltageFullRange ) ) -
       ( this.voltageSenseRange/2 );
-    return (float) voltage;
+    return voltage;
     
   }
 }
